Reinvented the structure of the request system and completed it (for movement' requests and transaction'). Made some general contracts modifications.
Created ParameterSettersExposer, which symbolizes the property of an object to expose certain methods (in particular, for requests, to expose parameter setters).
Created TransactionRequestFactory.
Created TransactionRequestParameters.
Extended ParameterSettersExposer with RequestBuilder, in fact, with the new system, every RequestBuilder must expose the setters for its parameters.
Consequently modified MovementRequest and TransactionRequest to make them implement getExposedMethods(...).
Now AbstractRequestFactory has the responsibility to set the parameters of the request to create this just after.
So MovementRequestFactory and TransactionRequestFactory have now the responsibility to create the specific builder and to retrieve the setters of that specific builder.
RequestFactory, to create a request, now requires a List of Objects for each parameter.
All other modifications are only of contracts.